requirements.txt
ccxt
pandas
ta
langchain
langchain-groq
python-dotenv
requests
langgraph